+++ src/App.css
@@ -174,3 +174,12 @@ main {
   width: 90%;
   max-width: 20rem;
 }
+
+.playlist-creation-input:focus {
+  outline: 1px solid gray;
+  padding: 0.3rem;
+}
+
+.float-right {
+  float: right;
+}
