+++ src/Components/PrivateRoute.jsx
@@ -1,6 +1,7 @@
-import { useAuth } from "./contexts/auth-context";
+import { useAuth } from "../contexts/auth-context";
 import { Route, useLocation } from "react-router";
 import { Navigate } from "react-router-dom";
+
 export function PrivateRoute({ path, element }) {
   const { token } = useAuth();
   const location = useLocation();
