+++ src/Components/Toast.jsx
@@ -1,6 +1,6 @@
 export const Toast = ({ message }) => {
   return (
-    <div className="snackbar" style={{ width: "17rem" }}>
+    <div className="snackbar snackbar-bottom-left" style={{ width: "17rem" }}>
       {message}
     </div>
   );
